SQL: implementation of the ORDER BY clause
Initial implementation of the SQL's ORDER BY clause. It supports
othe ASC and DESC sortings, but there are some limitations.

The sorting is only supported only if there is a matching index that
stores the key/value pairs already pre-sorted. Otherwise, an exception
is thrown.

The nested sortings are supported, but again only if the supporting
composite SORTED index is created.